Okey/requirements.txt
pre-commit==3.6.1
python-gitlab==4.4.0
python-dotenv==1.0.1
ansible==9.5.1
